1.調整金額可以顯示負數 2.調整餘額為0提示前往錢包 3.調整登入UI

+++ app/src/main/java/com/dae/stems_campus/data/model/ProfileModel.kt
@@ -13,7 +13,7 @@ class ProfileModel {
         val canUse: Boolean? = null,
         val isSuspended: Boolean? = null,
         val inDebt: Boolean? = null,
-        val debtAmount: Int? = null,
+        val debtAmount: Double? = null,
         val registrationSource: String? = null,
         val createdAt: String? = null,
         val updatedAt: String? = null,
